[04] Boot: Get it booting with Zig
Linker script is the same as the book. 'boot' is the Zig equivalent of
the book. 'kernel_main' is the same except it uses '@memset' from the
standard library.

Unlike the book when the kernel is run the stack pointer is 0x10 (16
bytes) from the initial stack pointer. This is because the Zig compiler
generates a function prologue for 'kernel_main'. If we comment out the
'@memset' call, Zig removes the function prologue and the stack pointer
is the same as the initial one shown in the disassembly.

While we're at it, bump the flake nixpkgs input to 25.05 (Zig 0.14.1).