[13] User Mode: Trigger page fault from user mode
The embedded user binary is now run in user mode by setting the
entrypoint address, setting user mode, then sret'ing to it. We trigger a
page fault by trying to change some kernel memory. It works!

[kern][err] PANIC: Unexpected trap scause=f, stval=80200000, user_pc=1019318

You can see that SCAUSE is 15, which corresponds to "Store/AMO page
fault". stval is the address we tried to modify (the start of kernel
memory). Finally, user_pc is 1019318. Inspecting the disassembly below,
we can see it matches up with our intentional illegal store.

   01019302 <user.main>:
   [...]
   ; /Users/fng/src/colonel/src/user.zig:18
   ;     const bad_ptr: *usize = @ptrFromInt(0x80200000);
    101930e:       sw      a1, -0xc(s0)
    1019312:       lui     a0, 0x1
    1019314:       addi    a0, a0, 0x234
   ; /Users/fng/src/colonel/src/user.zig:19
   ;     bad_ptr.* = 0x1234;
->  1019318:       sw      a0, 0x0(a1)
    101931a:       j       0x101931c <user.main+0x1a>